my-awesome-project/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
mypy_extensions==1.1.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pathspec==1.0.4
platformdirs==4.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
requests==2.32.5
ruff==0.15.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
